feature: Adds columns to KeySelector
refs #10

This adds each subkey's last write time, as well as its subkey counts,
to new cells in the `KeySelector` table widget. This is some
bare-minimum useful information to have available when navigating the
list. All three of the cells in the `KeySelector` table should be
sortable when we implement sorting as a part of #5.

This also adds some refactoring to the `key_selector` module. Lots of
logic that was aggregating inside of the `render` method has been moved
out into structs wrappers and implementations that implement the `From`
trait between the wrappers and the UI components.

Keys/Subkeys are now wrapped in structs that provide a `From` trait
implementation for converting to their respective `Table`/`Row` ratatui
UI components.